requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap4==2022.1
Django==4.2.4
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap4==23.2
django-crispy-forms==2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
fontawesomefree==6.4.2
phonenumbers==8.13.18
Pillow==10.0.0
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
shortuuid==1.0.11
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
